requirements.txt
Flask==2.0.1
python-dotenv==0.19.2
requests==2.26.0
Werkzeug==2.2.2
pydantic>=1.9.0,<2.0
supabase==1.0.3
gunicorn==20.1.0
openai==1.35.14
user-agents==2.2.0
stripe==2.60.0

# Add Speechify API library when available
